+++ config/generate/main.go
@@ -11,12 +11,13 @@ import (
 	"path/filepath"
 	"reflect"
 	"regexp"
+	"sort"
 	"strings"
 
 	"github.com/algolia/algoliasearch-client-go/v3/algolia/search"
 
 	"github.com/avenga/couper/config"
-	"github.com/avenga/couper/errors"
+	"github.com/avenga/couper/config/meta"
 )
 
 type entry struct {
@@ -49,20 +50,39 @@ func main() {
 	client := search.NewClient(searchAppID, os.Getenv(searchClientKey))
 	index := client.InitIndex(searchIndex)
 
+	filenameRegex := regexp.MustCompile(`(URL|JWT|OpenAPI|[a-z]+)`)
+	bracesRegex := regexp.MustCompile(`{([^}]*)}`)
+
+	attributesMap := map[string][]reflect.StructField{
+		"RequestHeadersAttributes":  newFields(&meta.RequestHeadersAttributes{}),
+		"ResponseHeadersAttributes": newFields(&meta.ResponseHeadersAttributes{}),
+		"FormParamsAttributes":      newFields(&meta.FormParamsAttributes{}),
+		"QueryParamsAttributes":     newFields(&meta.QueryParamsAttributes{}),
+		"LogFieldsAttribute":        newFields(&meta.LogFieldsAttribute{}),
+	}
+
 	for _, impl := range []interface{}{
 		&config.API{},
 		&config.Backend{},
 		&config.BasicAuth{},
 		&config.CORS{},
 		&config.Defaults{},
-		&config.Files{},
-		&config.Proxy{},
 		&config.Endpoint{},
+		&config.Files{},
 		&config.Health{},
+		&config.JWTSigningProfile{},
+		&config.OpenAPI{},
+		&config.Proxy{},
+		&config.Request{},
+		&config.Server{},
+		&config.Settings{},
+		&config.TokenRequest{},
+		&config.Websockets{},
 	} {
 		t := reflect.TypeOf(impl).Elem()
-		name := strings.TrimPrefix(strings.ToLower(fmt.Sprintf("%v", t)), "config.")
-		fileName := errors.TypeToSnake(impl)
+		name := reflect.TypeOf(impl).String()
+		name = strings.TrimPrefix(name, "*config.")
+		fileName := strings.ToLower(strings.Trim(filenameRegex.ReplaceAllString(name, "${1}_"), "_"))
 
 		result := entry{
 			Name: name,
@@ -80,7 +100,12 @@ func main() {
 		if ok {
 			it := reflect.TypeOf(inlineType.Inline()).Elem()
 			for i := 0; i < it.NumField(); i++ {
-				fields = append(fields, it.Field(i))
+				field := it.Field(i)
+				if _, ok := attributesMap[field.Name]; ok {
+					fields = append(fields, attributesMap[field.Name]...)
+				} else {
+					fields = append(fields, field)
+				}
 			}
 		}
 
@@ -107,13 +132,12 @@ func main() {
 				fieldDefault = "false"
 			} else if fieldDefault == "" && strings.HasPrefix(fieldType, "tuple ") {
 				fieldDefault = "[]"
-			} else if fieldType == "string" || fieldType == "duration" {
+			} else if fieldDefault != "" && (fieldType == "string" || fieldType == "duration") {
 				fieldDefault = `"` + fieldDefault + `"`
 			}
 
 			fieldDescription := field.Tag.Get("docs")
-			re := regexp.MustCompile(`{([^}]*)}`)
-			fieldDescription = re.ReplaceAllString(fieldDescription, "`${1}`")
+			fieldDescription = bracesRegex.ReplaceAllString(fieldDescription, "`${1}`")
 
 			a := attr{
 				Default:     fieldDefault,
@@ -123,6 +147,8 @@ func main() {
 			}
 			result.Attributes = append(result.Attributes, a)
 		}
+
+		sort.Sort(byName(result.Attributes))
 
 		b := &bytes.Buffer{}
 		enc := json.NewEncoder(b)
@@ -157,7 +183,7 @@ values: %s
 				continue
 			}
 
-			if line == "::" {
+			if skipMode && line == "::" {
 				skipMode = false
 				continue
 			}
@@ -178,10 +204,15 @@ values: %s
 `, b.String()))
 		}
 
-		_, err = file.WriteAt(fileBytes.Bytes(), 0)
+		size, err := file.WriteAt(fileBytes.Bytes(), 0)
 		if err != nil {
 			panic(err)
 		}
+		err = os.Truncate(file.Name(), int64(size))
+		if err != nil {
+			panic(err)
+		}
+
 		println("Attributes written: " + fileName)
 
 		if os.Getenv(searchClientKey) != "" {
@@ -192,4 +223,25 @@ values: %s
 			println("SearchIndex updated")
 		}
 	}
+}
+
+type byName []attr
+
+func (attributes byName) Len() int {
+	return len(attributes)
+}
+func (attributes byName) Swap(i, j int) {
+	attributes[i], attributes[j] = attributes[j], attributes[i]
+}
+func (attributes byName) Less(i, j int) bool {
+	return attributes[i].Name < attributes[j].Name
+}
+
+func newFields(impl interface{}) []reflect.StructField {
+	it := reflect.TypeOf(impl).Elem()
+	var fields []reflect.StructField
+	for i := 0; i < it.NumField(); i++ {
+		fields = append(fields, it.Field(i))
+	}
+	return fields
 }