+++ config/server.go
@@ -3,21 +3,23 @@ package config
 import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/gohcl"
+
+	"github.com/avenga/couper/config/meta"
 )
 
 var _ Inline = &Server{}
 
 // Server represents the <Server> object.
 type Server struct {
-	AccessControl        []string    `hcl:"access_control,optional"`
+	AccessControl        []string    `hcl:"access_control,optional" docs:"[access controls](../access-control) to protect the server. Inherited by nested blocks."`
 	APIs                 APIs        `hcl:"api,block"`
-	BasePath             string      `hcl:"base_path,optional"`
+	BasePath             string      `hcl:"base_path,optional" docs:"the path prefix for all requests"`
 	CORS                 *CORS       `hcl:"cors,block"`
-	DisableAccessControl []string    `hcl:"disable_access_control,optional"`
+	DisableAccessControl []string    `hcl:"disable_access_control,optional" docs:"disables access controls by name"`
 	Endpoints            Endpoints   `hcl:"endpoint,block"`
-	ErrorFile            string      `hcl:"error_file,optional"`
+	ErrorFile            string      `hcl:"error_file,optional" docs:"location of the error file template"`
 	Files                FilesBlocks `hcl:"files,block"`
-	Hosts                []string    `hcl:"hosts,optional"`
+	Hosts                []string    `hcl:"hosts,optional" docs:""`
 	Name                 string      `hcl:"name,label,optional"`
 	Remain               hcl.Body    `hcl:",remain"`
 	SPAs                 SPAs        `hcl:"spa,block"`
@@ -34,10 +36,8 @@ func (s Server) HCLBody() hcl.Body {
 // Inline implements the <Inline> interface.
 func (s Server) Inline() interface{} {
 	type Inline struct {
-		AddResponseHeaders map[string]string         `hcl:"add_response_headers,optional"`
-		DelResponseHeaders []string                  `hcl:"remove_response_headers,optional"`
-		SetResponseHeaders map[string]string         `hcl:"set_response_headers,optional"`
-		LogFields          map[string]hcl.Expression `hcl:"custom_log_fields,optional"`
+		meta.ResponseHeadersAttributes
+		meta.LogFieldsAttribute
 	}
 
 	return &Inline{}
@@ -51,6 +51,5 @@ func (s Server) Schema(inline bool) *hcl.BodySchema {
 	}
 
 	schema, _ := gohcl.ImpliedBodySchema(s.Inline())
-
-	return schema
+	return meta.MergeSchemas(schema, meta.ResponseHeadersAttributesSchema, meta.LogFieldsAttributeSchema)
 }